refactor(game): separate chunk generation from worker transport
